@@ -4,8 +4,9 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": ">=8.0",
+        "php": ">=8.1",
         "ext-mbstring": "*",
+        "cycle/entity-behavior": "^1.1",
         "doctrine/collections": "^1.6",
         "spiral/cycle-bridge": "^1.0",
         "spiral/framework": "^2.9",
