@@ -6,6 +6,7 @@ namespace App\Repository;
 
 use App\Database\Arena;
 use Cycle\ORM\Select\Repository;
+use DateTimeImmutable;
 
 class ArenaRepository extends Repository
 {
@@ -14,10 +15,17 @@ class ArenaRepository extends Repository
      */
     public function findActiveArenas(int $limit = null): array
     {
+        $past = new DateTimeImmutable('-1 minute');
         return $this
             ->select()
             ->load('characters')
-            ->where('active', true)
+            ->where([
+                'active' => true,
+                '@or' => [
+                    ['updatedAt' => null],
+                    ['updatedAt' => ['>' => $past]],
+                ],
+            ])
             ->limit($limit)
             ->fetchAll();
     }
